fix for mouse end frame

+++ src/preferences.h
@@ -401,7 +401,7 @@ typedef struct {
 
   boolean show_tooltips;
 
-  double volume; ///< audio volume level (for jack and pulse)
+  float volume; ///< audio volume level (for jack and pulse)
 
   boolean vj_mode; // optimise for VJing (in progress, experimental)
